+++ src/plugins/components/placeholder.js
@@ -1,0 +1,19 @@
+require('./placeholder.css')
+
+import {createElement as el} from 'react'
+
+
+export default function Placeholder (props) {
+  const {metadata} = props.initialBlock
+  const status = metadata ? metadata.status : ''
+
+  return el(
+    'div',
+    {className: 'Placeholder'},
+    el(
+      'p',
+      {className: 'Placeholder-status'},
+      status
+    )
+  )
+}