+++ CHANGES.md
@@ -1,7 +1,12 @@
+## 0.0.34 - 2016-01-31
+
+* [Demo implementation](./demo/demo.js) of Ed's `options.onShareUrl`
+* [Placeholder component](./src/plugins/components/placeholder.js) to show placeholder metadata.status for uploads and shares
+
 ## 0.0.33 - 2016-01-29
 
 * Updated to [ProseMirror 0.2.0](https://discuss.prosemirror.net/t/release-0-2-0/154/1)
 
 ## 0.0.32 - 2016-01-29
 
-* [Demo implementation](./demo/demo.js) of Ed's `options.onShareFile` 
+* [Demo implementation](./demo/demo.js) of Ed's `options.onShareFile`